Add repl config for sqlcmd.
Because a database user typically works on many different servers and
databases I don't see a way to handle this except to create a config for
each server that you want to connect too. This is within reason as you
set it up once and can create a new sqlcmd REPL for the given server. I
have included a local host config that points to a database on the
localhost. Anyone wanting to use this will need to update the config to
match there server and database name.