Attempted fix for windows "handle is invalid" error.
I don't know if this is wise, but this is the smallest patch I could make
that fixes the "handle is invalid" bug.  I do not fully understand why
the version of CreateProcessW in _winapi works and the one in winprocess
does not.  There have been a lot of changes to the Popen class, and the
version of _execute_child in SublimeREPL has diverged a bit from the one
in the subprocess module, so maybe there are some bad subtle assumptions.

This probably won't work in Sublime 2.  I could look into accomodating
that as well (probably just need to use _subprocess).  However, I'm not
sure this is really the best approach.