Change Syntax to Plain for REPL output
This changes the syntax highlighting from Python to Plain Text in the case of Run file or (run file under debugger). Before the fix the output written to the *REPL* window, which has it's view changed to look just like a terminal instead of IDLE, was colored oddly according to Python syntax, (which confused and frustrated my students making interactive text-based adventure games). 

In fact, it would be awesome if it were easier to set our own syntax files for the REPL output so we could make one that matched the interactive output of our little games.